% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_scroll.R
\name{fixed_areas_bundle}
\alias{fixed_areas_bundle}
\title{Creates a bundle of fixed areas mappings}
\usage{
fixed_areas_bundle(...)
}
\arguments{
\item{...}{A varying number of pairs of coordinate vectors. The first of these
pairs should always represent an area on the screen, and the second one an
area on the unscrolled image. Each vector should take the form c(top_left_x,
top_left_y, bottom_right_x, bottom_right_y)}
}
\value{
A bundle of fixed areas mappings, to be put in a list
}
\description{
Puts together several fixed areas mappings to make them all
(de)activated by the same rule
}
\examples{
\dontrun{
top_fixed_area_screen <- c(0, 89, 1919, 276)
top_fixed_area_image <- c(0, 0, 1919, 187)
right_fixed_area_screen <- c(1632, 277, 1919, 936)
right_fixed_area_image <- c(1632, 188, 1919, 847)
area_bundle <- fixed_areas_bundle(top_fixed_area_screen,
                                  top_fixed_area_image,
                                  right_fixed_area_screen,
                                  right_fixed_area_image)
fixed_areas <- list(area_bundle)
}
}
