% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_scroll.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Creates generic heatmap}
\usage{
generate_heatmap(data, heatmap_image)
}
\arguments{
\item{data}{The dataset output by the \code{\link{eye_scroll_correct}} function}

\item{heatmap_image}{The unscrolled webpage image on which to apply the heatmap}
}
\value{
A plot with the image and the heatmap
}
\description{
Creates a heatmap based on the dataset and the unscrolled webpage image
}
\examples{
\dontrun{
img <- readPNG("test.png")
test_data <- eye_scroll_correct(fixed_areas = test_fixed_areas,
    rules = test_rules, calibration = test_calibration)
generate_heatmap(data = test_data, heatmap_image = img)
}
}
