% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracker_correct_gaze.R
\name{rule_before_scrolling}
\alias{rule_before_scrolling}
\title{rule_before_scrolling}
\usage{
rule_before_scrolling(timestamp, event, x, y, array_anchors, flag, scroll)
}
\arguments{
\item{timestamp}{The timestamp of the current line in the .csv file}

\item{event}{The event column from this line}

\item{x}{The x parameter (eye gaze) from this line}

\item{y}{The y parameter (eye gaze) from this line}

\item{array_anchors}{The array of anchors linked to this rule}

\item{flag}{A boolean that says if this rule is TRUE or FALSE at the moment}

\item{scroll}{The total amount of pixels that have been scrolled down from
the top of the website at the moment}
}
\value{
A boolean saying if this rule should now be enforced or not
}
\description{
Helper function to enforce anchor rules (this one checks
    if the total amount of pixels that have been scrolled down from the top
    of the website is inferior to 30). NB: you should never
    have to call this function yourself, but you can use it as a basis to
    create your own rule. All arguments are automatically passed to any
    rule function by eye_tracker_gaze_scroll
}
\examples{
rules <- list(rule_before_scrolling, rule_after_scrolling, rule_true)
}
