% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_scroll.R
\name{scroll_calibration_auto}
\alias{scroll_calibration_auto}
\title{Automatic calibration}
\usage{
scroll_calibration_auto(calibration_image, scroll_pixels)
}
\arguments{
\item{calibration_image}{The screenshot, loaded as a large array (e.g. using readPNG)}

\item{scroll_pixels}{The amount of pixels being scrolled by each mouse scroll
(depends on the browser). Output by the calibration webpage}
}
\value{
A list representing the calibration to the screen & browser used for the experiment
}
\description{
Creates a calibration list for use in the eye_scroll_correct
    function. Uses a screenshot of a full screen including the calibration webpage
}
\examples{
\dontrun{
library(png)
img <- readPNG("calibrate.png")
calibration <- scroll_calibration_auto(calibration_image = img, scroll_pixels = 100)
}
}
