% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracker_fixation_scroll.R
\name{scroll_calibration_manual}
\alias{scroll_calibration_manual}
\title{scroll_calibration_manual}
\usage{
scroll_calibration_manual(
  screen_width,
  screen_height,
  top_left_x,
  top_left_y,
  bottom_right_x,
  bottom_right_y,
  scroll_pixels
)
}
\arguments{
\item{screen_width}{Resolution width of the screen on which the experiment is conducted (in pixels). E.g. 1920}

\item{screen_height}{Resolution height of the screen on which the experiment is conducted (in pixels). E.g. 1080}

\item{top_left_x}{x coordinate of the top left pixel of the browser viewing area}

\item{top_left_y}{y coordinate of the same pixel}

\item{bottom_right_x}{x coordinate of the bottom right pixel of the browser viewing area}

\item{bottom_right_y}{y coordinate of the same pixel}

\item{scroll_pixels}{The amount of pixels being scrolled by each mouse scroll
(depends on the browser). Output by the calibration webpage}
}
\value{
A list representing the calibration to the screen & browser used for the experiment
}
\description{
Creates a calibration list for use in the eye_tracker_fixation_scroll
    function. Uses values taken by hand
}
\examples{
calibration <- scroll_calibration_manual(1920, 1080, 88, 0, 40, 0, 100)
}
