% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracker_correct_gaze.R
\name{eye_tracker_gaze_scroll}
\alias{eye_tracker_gaze_scroll}
\title{eye_tracker_gaze_scroll}
\usage{
eye_tracker_gaze_scroll(
  file_name = "mouse events and fixation coordinates.csv",
  time_shift = 388,
  timestamp_start = 53753,
  image_height = 6655,
  image_width = 1920,
  timestamp_stop = 113579,
  starting_scroll = 0,
  output_file = "eye_tracker_gaze_corrected.csv",
  anchors = list(array(c(c(0, 0, 1902, 95), c(0, 0, 1902, 95), c(1607, 96, 1902, 905),
    c(1607, 96, 1902, 905)), dim = c(4, 2, 2)), array(c(c(0, 0, 1902, 63), c(0, 32, 1902,
    95), c(1607, 63, 1902, 900), c(1607, 63, 1902, 900)), dim = c(4, 2, 2)),
    array(c(c(1607, 906, 1920, 1080), c(1607, 6481, 1920, 6655)), dim = c(4, 2, 1))),
  rules = list(rule_before_scrolling, rule_after_scrolling, rule_true),
  outside_image_is_na = TRUE,
  na.rm = TRUE,
  calibration
)
}
\arguments{
\item{file_name}{The name of the .csv file with the coordinates & input events}

\item{time_shift}{A time shift parameter to synchronize the .csv file with a video recording (e.g. if the recording started after the start of the .csv)}

\item{timestamp_start}{The starting timestamp from the .csv file at which the participant was watching the scrollable (AFTER time_shift applied)}

\item{image_height}{The total height of the image (in pixels)}

\item{image_width}{The total width of the image (in pixels)}

\item{timestamp_stop}{The final timestamp from the .csv file at which the participant was watching the scrollable (AFTER time_shift applied)}

\item{starting_scroll}{[Optional] If the participant did not start watching the webpage from the top, you can indicate the y coordinate at which he started here (in pixels)}

\item{output_file}{The name of the output .csv file. If empty, will just return the data without creating a new file}

\item{anchors}{A list of potentially immovable areas inside the scrollable (e.g. anchored menus in a website) - see manual for a correct usage}

\item{rules}{A list of functions that can act as rules to activate/deactivate immovables areas in the scrollable (e.g. a menu that disappears after X pixels have been scrolled) - See manual for a correct}

\item{outside_image_is_na}{Indicates if values outside the AOI (e.g. the windows bar) should be set to NA or kept in the file/dataset. If set to False, coordinates above/before the AOI will become negative.}

\item{na.rm}{Indicates if lines with NA y values should be dropped in the final file/dataset}

\item{calibration}{A calibration list (see the scroll_calibration function)}
}
\value{
The Dataset with corrected x and y's
}
\description{
corrects the eye-tracking coordinate data to fit in a webpage
    scrolled vertically by the participant
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
