% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracker_fixation_scroll.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{generate_heatmap}
\usage{
generate_heatmap(data, heatmap_image)
}
\arguments{
\item{data}{The dataset output by the eye_tracker_fixation_scroll}

\item{heatmap_image}{An image on which to apply the heatmap}
}
\value{
A plot with the image and the heatmap
}
\description{
Creates a heatmap based on the dataset
}
\examples{
\dontrun{
img <- readPNG("test.png")
test_data <- eye_tracker_fixation_scroll(anchors = test_anchors,
    rules = test_rules, calibration = test_calibration)
generate_heatmap(data = test_data, heatmap_image = img)
}
}
