% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_scroll.R
\name{rule_before_scrolling}
\alias{rule_before_scrolling}
\title{Generic rule to activate a bundle before a certain amount of scrolled
pixels}
\usage{
rule_before_scrolling(data_line, fixed_areas_bundle, flag, scroll)
}
\arguments{
\item{data_line}{The current line in the .csv file. Includes all the original
columns, along with a new "Timestamp.Shifted" column (Timestamp - time_shit),
along with "Corrected.X" and "Corrected.Y" columns, which are "Fixation.X"
and "Fixation.Y" respectively shifted by top_left_x and top_left_y of the
calibration stage}

\item{fixed_areas_bundle}{The array of fixed areas linked to this rule}

\item{flag}{A boolean that says if this rule is TRUE or FALSE at the moment}

\item{scroll}{The total amount of pixels that have been scrolled down from
the top of the website at the moment}
}
\value{
A boolean saying if this rule should now be enforced or not
}
\description{
Helper function to enforce fixed area rules (this one checks
    if the total amount of pixels that have been scrolled down from the top
    of the website is inferior to 30). NB: you should probably not
    have to call this function yourself, but you can use it as a basis to
    create your own rule. All arguments are automatically passed to any
    rule function by eye_scroll_correct
}
\examples{
rule_before_scrolling <- function (data_line, fixed_areas_bundle, flag, scroll)
{
  if (scroll < 30)
  {
    return (TRUE)
  }
  else
  {
    return(FALSE)
  }
}
rules <- list(rule_before_scrolling, rule_after_scrolling, rule_true)
}
