% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracker_fixation_scroll.R
\name{scroll_calibration}
\alias{scroll_calibration}
\title{scroll_calibration}
\usage{
scroll_calibration(
  screen_width,
  screen_height,
  shift_top,
  shift_left,
  shift_bottom,
  shift_right,
  scroll_pixels
)
}
\arguments{
\item{screen_width}{Resolution width of the screen on which the experiment is conducted (in pixels). E.g. 1920}

\item{screen_height}{Resolution height of the screen on which the experiment is conducted (in pixels). E.g. 1080}

\item{shift_top}{The number of pixels above the area of interest on the screen (and therefore, outside of the image)}

\item{shift_left}{Same, but on the left of the AOI}

\item{shift_bottom}{Same, but below the AOI}

\item{shift_right}{Same, but on the right of the AOI}

\item{scroll_pixels}{The amount of pixels being scrolled by each mouse scroll
(depends on the browser)}
}
\value{
A list representing the calibration to the screen & browser used for the experiment
}
\description{
Creates a calibration list for use in the eye_tracker_fixation_scroll
    function. Can be filled in by taking measurements by hand, or using the
    calibration tool.
}
\examples{
calibration <- scroll_calibration(1920, 1080, 88, 0, 40, 0, 100)
}
